Revert "Agesa bootblock"
