+++ src/drivers/amd/agesa/bootblock.c
@@ -2,20 +2,14 @@
 
 #include <bootblock_common.h>
 #include <halt.h>
-#include <string.h>
-#include <acpi/acpi.h>
 #include <amdblocks/amd_pci_mmconf.h>
+#include <amdblocks/biosram.h>
 #include <arch/bootblock.h>
-#include <arch/cpu.h>
 #include <cpu/amd/msr.h>
 #include <cpu/x86/mtrr.h>
 #include <cpu/x86/lapic.h>
-#include <northbridge/amd/agesa/agesa_helper.h>
-#include <northbridge/amd/agesa/state_machine.h>
 
 #define EARLY_VMTRR_FLASH 6
-
-void __weak board_BeforeAgesa(struct sysinfo *cb) { }
 
 static void set_early_mtrrs(void)
 {
@@ -27,39 +21,6 @@ static void set_early_mtrrs(void)
 void bootblock_soc_early_init(void)
 {
 	bootblock_early_southbridge_init();
-}
-
-static void fill_sysinfo(struct sysinfo *cb)
-{
-	memset(cb, 0, sizeof(*cb));
-	cb->s3resume = acpi_is_wakeup_s3();
-
-	agesa_set_interface(cb);
-}
-
-void bootblock_soc_init(void)
-{
-	struct sysinfo bootblock_state;
-	struct sysinfo *cb = &bootblock_state;
-	u8 initial_apic_id = (u8) (cpuid_ebx(1) >> 24);
-
-	fill_sysinfo(cb);
-
-	if (initial_apic_id == 0) {
-		board_BeforeAgesa(cb);
-		console_init();
-	}
-
-	printk(BIOS_DEBUG, "APIC %02d: CPU Family_Model = %08x\n",
-		initial_apic_id, cpuid_eax(1));
-
-	agesa_execute_state(cb, AMD_INIT_RESET);
-
-	agesa_execute_state(cb, AMD_INIT_EARLY);
-
-	/* APs shall not get here */
-	if (initial_apic_id != 0)
-		halt();
 }
 
 asmlinkage void bootblock_c_entry(uint64_t base_timestamp)
@@ -81,5 +42,7 @@ asmlinkage void ap_bootblock_c_entry(void)
 	if (CONFIG(UDELAY_LAPIC))
 		enable_lapic();
 
-	bootblock_soc_init();
+	void (*ap_romstage_entry)(void) = get_ap_entry_ptr();
+	ap_romstage_entry(); /* execution does not return */
+	halt();
 }
