+++ src/drivers/amd/agesa/romstage.c
@@ -15,6 +15,8 @@
 #include <northbridge/amd/agesa/agesa_helper.h>
 #include <northbridge/amd/agesa/state_machine.h>
 
+void __weak board_BeforeAgesa(struct sysinfo *cb) { }
+
 static void fill_sysinfo(struct sysinfo *cb)
 {
 	memset(cb, 0, sizeof(*cb));
@@ -22,6 +24,11 @@ static void fill_sysinfo(struct sysinfo *cb)
 
 	agesa_set_interface(cb);
 }
+
+/* APs will enter directly here from bootblock, bypassing verstage
+ * and potential fallback / normal bootflow detection.
+ */
+static void ap_romstage_main(void);
 
 static void romstage_main(void)
 {
@@ -33,8 +40,23 @@ static void romstage_main(void)
 
 	fill_sysinfo(cb);
 
-	if (initial_apic_id == 0)
+	if (initial_apic_id == 0) {
+
+		timestamp_add_now(TS_START_ROMSTAGE);
+
+		board_BeforeAgesa(cb);
+
 		console_init();
+	}
+
+	printk(BIOS_DEBUG, "APIC %02d: CPU Family_Model = %08x\n",
+		initial_apic_id, cpuid_eax(1));
+
+	set_ap_entry_ptr(ap_romstage_main);
+
+	agesa_execute_state(cb, AMD_INIT_RESET);
+
+	agesa_execute_state(cb, AMD_INIT_EARLY);
 
 	timestamp_add_now(TS_BEFORE_INITRAM);
 
@@ -64,6 +86,21 @@ static void romstage_main(void)
 
 	run_postcar_phase(&pcf);
 	/* We do not return. */
+}
+
+static void ap_romstage_main(void)
+{
+	struct sysinfo romstage_state;
+	struct sysinfo *cb = &romstage_state;
+
+	fill_sysinfo(cb);
+
+	agesa_execute_state(cb, AMD_INIT_RESET);
+
+	agesa_execute_state(cb, AMD_INIT_EARLY);
+
+	/* Not reached. */
+	halt();
 }
 
 asmlinkage void car_stage_entry(void)
