+++ src/mainboard/pcengines/apu2/Kconfig
@@ -119,18 +119,19 @@ config FMDFILE
 
 config CBFS_SIZE
 	hex
+	default 0x0020C000 if VBOOT
 	default 0x00800000
 
 config VBOOT
-	select VBOOT_VBNV_FLASH
+	select VBOOT_VBNV_CMOS
 	select VBOOT_NO_BOARD_SUPPORT
 	select GBB_FLAG_DISABLE_LID_SHUTDOWN
 	select GBB_FLAG_DISABLE_PD_SOFTWARE_SYNC
 	select GBB_FLAG_DISABLE_EC_SOFTWARE_SYNC
 	select GBB_FLAG_DISABLE_FWMP
 	select RTC
-	select VBOOT_STARTS_IN_BOOTBLOCK
-	select VBOOT_SEPARATE_VERSTAGE
+	select VBOOT_STARTS_IN_ROMSTAGE
+	select AMDFW_OUTSIDE_CBFS
 
 config VBOOT_MEASURED_BOOT_RUNTIME_DATA
 	string
@@ -138,11 +139,11 @@ config VBOOT_MEASURED_BOOT_RUNTIME_DATA
 
 config VBOOT_VBNV_OFFSET
 	hex
-	default 0x2D8 if VBOOT_VBNV_CMOS
+	default 0x2D8 if VBOOT
 
 config RO_REGION_ONLY
 	string
-	default "AGESA apu/amdfw" if VBOOT
+	default "spd.bin AGESA" if VBOOT
 
 config AGESA_USE_1_0_0_4_HEADER
 	bool
