+++ src/mainboard/pcengines/apu2/Makefile.inc
@@ -3,12 +3,6 @@
 bootblock-y += pmutil.c
 bootblock-y += bootblock.c
 bootblock-y += bios_knobs.c
-bootblock-y += BiosCallOuts.c
-bootblock-y += OemCustomize.c
-bootblock-y += gpio_ftns.c
-
-verstage-y += pmutil.c
-verstage-y += bios_knobs.c
 
 romstage-y += bios_knobs.c
 romstage-y += BiosCallOuts.c
