+++ src/mainboard/pcengines/apu2/vboot-rw-ab.fmd
@@ -1,32 +1,28 @@
 FLASH 8M {
 	SI_BIOS@0x0 0x800000 {
-		RW_MISC@0x0 0x10000 {
-			RW_SHARED@0x0 0x4000 {
-				SHARED_DATA@0x0 0x2000
-				VBLOCK_DEV@0x2000 0x2000
-			}
-			RW_VPD(PRESERVE)@0x4000 0x2000
-			RW_NVRAM(PRESERVE)@0x6000 0x6000
-			RW_ELOG(PRESERVE)@0xc000 0x4000
+		MISC_RW@0x0 0xa0000 {
+			RW_UNUSED@0x0 0x1c000
+			RW_VPD@0x1c000 0x4000
+			PSPDIR@0x20000 0x80000
 		}
-		RW_SECTION_A@0x10000 0x1f8000 {
+		RW_SECTION_A@0xa0000 0x230000 {
 			VBLOCK_A@0x0 0x10000
-			FW_MAIN_A(CBFS)@0x10000 0x1e7fc0
-			RW_FWID_A@0x1f7fc0 0x40
+			FW_MAIN_A(CBFS)@0x10000 0x21ffc0
+			RW_FWID_A@0x22ffc0 0x40
 		}
-		RW_SECTION_B@0x208000 0x1f8000 {
+		RW_SECTION_B@0x2d0000 0x230000 {
 			VBLOCK_B@0x0 0x10000
-			FW_MAIN_B(CBFS)@0x10000 0x1e7fc0
-			RW_FWID_B@0x1f7fc0 0x40
+			FW_MAIN_B(CBFS)@0x10000 0x21ffc0
+			RW_FWID_B@0x22ffc0 0x40
 		}
-		WP_RO@0x400000 0x400000{
+		WP_RO@0x500000 0x300000{
 			RO_VPD@0x0 0x4000
-			RO_SECTION@0x4000 0x3fc000{
+			RO_SECTION@0x4000 0x2fc000{
 				FMAP@0x0 0x800
 				RO_FRID@0x800 0x40
 				RO_FRID_PAD@0x840 0x7c0
 				GBB@0x1000 0xef000
-				COREBOOT(CBFS)@0xf0000 0x30c000
+				COREBOOT(CBFS)@0xf0000 0x20c000
 			}
 		}
 	}
