+++ src/northbridge/amd/agesa/family15tn/Makefile.inc
@@ -5,6 +5,5 @@ romstage-y += dimmSpd.c
 ramstage-y += iommu.c
 ramstage-y += northbridge.c
 
-bootblock-y += state_machine.c
 romstage-y += state_machine.c
 ramstage-y += state_machine.c
