+++ src/northbridge/amd/agesa/family16kb/Makefile.inc
@@ -4,6 +4,5 @@ romstage-y += dimmSpd.c
 
 ramstage-y += northbridge.c
 
-bootblock-y += state_machine.c
 romstage-y += state_machine.c
 ramstage-y += state_machine.c
