+++ src/northbridge/amd/pi/00730F01/Makefile.inc
@@ -5,6 +5,5 @@ romstage-y += dimmSpd.c
 ramstage-y += northbridge.c
 ramstage-y += iommu.c
 
-bootblock-y += state_machine.c
 romstage-y += state_machine.c
 ramstage-y += state_machine.c
