+++ src/southbridge/amd/pi/hudson/Kconfig
@@ -19,7 +19,6 @@ config SOUTHBRIDGE_SPECIFIC_OPTIONS # dummy
 	select HAVE_CF9_RESET_PREPARE
 	select SOC_AMD_COMMON
 	select SOC_AMD_COMMON_BLOCK_ACPIMMIO
-	select BOOT_DEVICE_SUPPORTS_WRITES
 
 config EHCI_BAR
 	hex
